// 因外链界面tablinkName相同时，只激活一个tab,所以外链界面用linkName作为key存入session,避免因勾选的id不同存入多个，导致关闭当前tab时无法清除存入的多个

+++ src/__config__/mutations/version_1.4/global.mutations.js
@@ -388,9 +388,13 @@ export default {
 
 
     // 清除配置界面提供给定制界面的参数信息
-    const { customizedModuleId, linkModuleId } = router.currentRoute.params;
-    deleteFromSessionObject('customizeMessage', customizedModuleId);
-    deleteFromSessionObject('customizeMessage', linkModuleId);
+    // const { customizedModuleId } = router.currentRoute.params;
+    if (tab.keepAliveModuleName) {
+      const customizedModuleId = tab.keepAliveModuleName.split('.')[2];
+      deleteFromSessionObject('customizeMessage', customizedModuleId);
+    }
+   
+    deleteFromSessionObject('customizeMessage', tab.tableName);
 
 
     // 清除当前关闭的表单设置的跳转到标准列表表单默认值;
